[refactor] 불필요 코드 제거 및 리팩토링 #47
- LoginService.java : 불필요 코드 제거하고 토큰을 빌드하는 메서드 위치 옮겼습니다.
