+++ BE/sidedish11/src/main/java/dev/codesquad/java/sidedish11/controller/LoginController.java
@@ -1,6 +1,5 @@
 package dev.codesquad.java.sidedish11.controller;
 
-import dev.codesquad.java.sidedish11.jwt.TokenService;
 import dev.codesquad.java.sidedish11.oauth.Github;
 import dev.codesquad.java.sidedish11.oauth.GithubUser;
 import dev.codesquad.java.sidedish11.service.LoginService;
@@ -24,9 +23,6 @@ public class LoginController {
     @Autowired
     private LoginService loginService;
 
-    @Autowired
-    private TokenService tokenService;
-
     @GetMapping("/login")
     public ResponseEntity login(HttpServletResponse response) {
         response.setHeader(HEADER_LOCATION, LOGIN_REQUEST_URL);
@@ -37,9 +33,9 @@ public class LoginController {
     public ResponseEntity oauth(@PathParam("code") String code, HttpServletResponse response) {
         Github github = loginService.requestAccessToken(code);
         GithubUser githubUser = loginService.requestUserId(github.getAuthorization());
-        response.setHeader(HEADER_LOCATION, SERVER_URL);
+        String jwt = loginService.buildToken(githubUser);
 
-        String jwt = tokenService.buildToken(githubUser);
+        response.setHeader(HEADER_LOCATION, SERVER_URL);
         response.setHeader(AUTHORIZATION, jwt);
         response.setHeader(GITHUB_USER_ID, githubUser.getUserId());
         return new ResponseEntity(githubUser, HttpStatus.FOUND);
