+++ BE/sidedish11/src/main/resources/application.properties
@@ -5,23 +5,23 @@ spring.mvc.hiddenmethod.filter.enabled=true
 logging.level.sql=debug;
 logging.level.dev.codesquad.java.sidedish11=debug
 
-# ec2 mySQL 연결
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/sidedish11
-spring.datasource.username=henry
-spring.datasource.password=henry
-spring.datasource.initialization-mode=always
-spring.datasource.data=classpath:schema.sql, classpath:data.sql
-#spring.datasource.initialization-mode=embedded
+## ec2 mySQL 연결
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#spring.datasource.url=jdbc:mysql://localhost:3306/sidedish11
+#spring.datasource.username=henry
+#spring.datasource.password=henry
+#spring.datasource.initialization-mode=always
+#spring.datasource.data=classpath:schema.sql, classpath:data.sql
+##spring.datasource.initialization-mode=embedded
 
-## DB Connection 설정 (로컬 DB 테스트용)
-#spring.datasource.url=jdbc:h2:mem://localhost/~/sidedish-11;DB_CLOSE_ON_EXIT=FALSE
-#spring.datasource.driverClassName=org.h2.Driver
-#spring.datasource.username=sa
-#spring.datasource.password=
-#
-#spring.h2.console.enabled=true
-#spring.h2.console.path=/h2-console
+# DB Connection 설정 (로컬 DB 테스트용)
+spring.datasource.url=jdbc:h2:mem://localhost/~/sidedish-11;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
 
 # application.properties
 spring.profiles.active=secret
